requirements.txt
PyMuPDF>=1.23.0
python-docx>=1.1.0
deep-translator>=1.11.4
Pillow>=10.0.0
pytesseract>=0.3.10

